+++ app/datatables/category_results_datatable.rb
@@ -2,12 +2,13 @@ class CategoryResultsDatatable < IndexDatatable
   def initialize(*)
     super
     
-    columns([:competition_name, :competition_class, :competition_type, :category, :season,
-             :ranking, :skater_name, :nation, :points,
-             :short_ranking,
-             :short_tss, :short_tes, :short_pcs, :short_deductions, :short_base_value,
-             :free_ranking,
-             :free_tss, :free_tes, :free_pcs, :free_deductions, :free_base_value,
+    columns([
+              :competition_name, :competition_class, :competition_type, :category, :season,
+              :ranking, :skater_name, :nation, :points,
+              :short_ranking,
+              :short_tss, :short_tes, :short_pcs, :short_deductions, :short_base_value,
+              :free_ranking,
+              :free_tss, :free_tes, :free_pcs, :free_deductions, :free_base_value,
             ])
     
     columns.sources = {